Bring skipLayoutIdsFilter back as a README example.
`skipLayoutIdsFilter` was removed in a previous commit, with the justification that
I think `testTag` is better-suited for this purpose, `layoutId` is really only
meant for communicating with a particular custom parent layout. But I used the
implementation for a new example in the README of how to write a custom filter.